fixed very lurky context bug in node vs webkit execution context not reall y seen when reloading the wb when developing. ugh. affected filesync and upload
